Extract executor (#2733)
Extracts (one of the two...) `PowdrExecutor` into its own file. I just
moved the code (the only change is that I made the fields public for
now, to minimize code change).

I think the next steps should be:
- Make it do most of what `generate_air_proof_input` does as well
- Refactor (so the behavior is more encapsulated)
- Re-use it for witness generation of the PlonK chip
